+++ internal/licensing/license_test.go
@@ -1,4 +1,4 @@
-package core
+package licensing
 
 import "testing"